langgraph-backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sse-starlette==2.1.0
langgraph>=0.6.3
langchain>=0.3.27
langchain-community>=0.3.14
langchain-anthropic>=0.3.0
langchain-mcp-adapters>=0.1.9
sqlalchemy==2.0.36
alembic==1.13.1
aiosqlite==0.20.0
python-dotenv==1.0.1
pydantic>=2.7.4
pydantic-settings>=2.3.3
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
redis==5.0.5
prometheus-client==0.20.0
python-multipart==0.0.9
# PostgreSQL and vector dependencies for memory management
asyncpg==0.29.0
numpy>=1.24.0
openai>=1.51.0
psycopg2-binary==2.9.9
# Native LangChain/LangGraph memory integration
langchain-openai>=0.3.28
langchain-postgres>=0.0.15
pgvector>=0.3.6
tiktoken>=0.10.0
# Google Analytics Data API for GA4 integration
google-analytics-data>=0.18.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0